LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/the-canvas-state/2d.state.saverestore.stackdepth.worker-expected.txt


PASS save()/restore() stack depth is not unreasonably limited
